Revamp code actions
