+++ language-server/modules/langserver-core/src/main/java/org/ballerinalang/langserver/codeaction/providers/AddAllDocumentationCodeAction.java
@@ -16,16 +16,19 @@
 package org.ballerinalang.langserver.codeaction.providers;
 
 import org.ballerinalang.annotation.JavaSPIService;
+import org.ballerinalang.langserver.command.executors.AddAllDocumentationExecutor;
+import org.ballerinalang.langserver.common.constants.CommandConstants;
 import org.ballerinalang.langserver.commons.LSContext;
 import org.ballerinalang.langserver.commons.codeaction.CodeActionNodeType;
+import org.ballerinalang.langserver.commons.command.CommandArgument;
 import org.ballerinalang.langserver.compiler.DocumentServiceKeys;
 import org.eclipse.lsp4j.CodeAction;
+import org.eclipse.lsp4j.Command;
 
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
-
-import static org.ballerinalang.langserver.command.CommandUtil.getAllDocGenerationCommand;
 
 /**
  * Code Action provider for adding all documentation for top level items.
@@ -49,6 +52,12 @@ public class AddAllDocumentationCodeAction extends AbstractCodeActionProvider {
     @Override
     public List<CodeAction> getCodeActions(CodeActionNodeType nodeType, LSContext lsContext,
                                            List<org.eclipse.lsp4j.Diagnostic> diagnostics) {
-        return Collections.singletonList(getAllDocGenerationCommand(lsContext.get(DocumentServiceKeys.FILE_URI_KEY)));
+        String docUri = lsContext.get(DocumentServiceKeys.FILE_URI_KEY);
+        CommandArgument docUriArg = new CommandArgument(CommandConstants.ARG_KEY_DOC_URI, docUri);
+        List<Object> args = new ArrayList<>(Collections.singletonList(docUriArg));
+
+        CodeAction action = new CodeAction(CommandConstants.ADD_ALL_DOC_TITLE);
+        action.setCommand(new Command(CommandConstants.ADD_ALL_DOC_TITLE, AddAllDocumentationExecutor.COMMAND, args));
+        return Collections.singletonList(action);
     }
 }
