+++ language-server/modules/langserver-core/src/main/java/org/ballerinalang/langserver/codeaction/providers/CreateFunctionCodeAction.java
@@ -1,0 +1,130 @@
+/*
+ * Copyright (c) 2019, WSO2 Inc. (http://wso2.com) All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.ballerinalang.langserver.codeaction.providers;
+
+import org.ballerinalang.annotation.JavaSPIService;
+import org.ballerinalang.langserver.command.CommandUtil;
+import org.ballerinalang.langserver.command.executors.CreateFunctionExecutor;
+import org.ballerinalang.langserver.common.constants.CommandConstants;
+import org.ballerinalang.langserver.common.utils.CommonUtil;
+import org.ballerinalang.langserver.commons.LSContext;
+import org.ballerinalang.langserver.commons.codeaction.CodeActionKeys;
+import org.ballerinalang.langserver.commons.codeaction.CodeActionNodeType;
+import org.ballerinalang.langserver.commons.command.CommandArgument;
+import org.ballerinalang.langserver.commons.workspace.LSDocumentIdentifier;
+import org.ballerinalang.langserver.commons.workspace.WorkspaceDocumentException;
+import org.ballerinalang.langserver.commons.workspace.WorkspaceDocumentManager;
+import org.ballerinalang.langserver.compiler.DocumentServiceKeys;
+import org.ballerinalang.langserver.compiler.exception.CompilationFailedException;
+import org.ballerinalang.model.elements.PackageID;
+import org.eclipse.lsp4j.CodeAction;
+import org.eclipse.lsp4j.CodeActionKind;
+import org.eclipse.lsp4j.Command;
+import org.eclipse.lsp4j.Diagnostic;
+import org.eclipse.lsp4j.Position;
+import org.wso2.ballerinalang.compiler.tree.BLangPackage;
+import org.wso2.ballerinalang.compiler.tree.expressions.BLangInvocation;
+
+import java.nio.file.Path;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Optional;
+import java.util.regex.Matcher;
+
+/**
+ * Code Action provider for importing a package.
+ *
+ * @since 1.2.0
+ */
+@JavaSPIService("org.ballerinalang.langserver.commons.codeaction.spi.LSCodeActionProvider")
+public class CreateFunctionCodeAction extends AbstractCodeActionProvider {
+    private static final String UNDEFINED_FUNCTION = "undefined function";
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public List<CodeAction> getCodeActions(CodeActionNodeType nodeType, LSContext lsContext,
+                                           List<Diagnostic> diagnostics) {
+        WorkspaceDocumentManager documentManager = lsContext.get(CodeActionKeys.DOCUMENT_MANAGER_KEY);
+        Optional<Path> filePath = CommonUtil.getPathFromURI(lsContext.get(CodeActionKeys.FILE_URI_KEY));
+        LSDocumentIdentifier document = null;
+        try {
+            document = documentManager.getLSDocument(filePath.get());
+        } catch (WorkspaceDocumentException e) {
+            // ignore
+        }
+        List<CodeAction> actions = new ArrayList<>();
+
+        if (document == null) {
+            return actions;
+        }
+        for (Diagnostic diagnostic : diagnostics) {
+            if (diagnostic.getMessage().startsWith(UNDEFINED_FUNCTION)) {
+                CodeAction codeAction = getFunctionImportCommand(document, diagnostic, lsContext);
+                if (codeAction != null) {
+                    actions.add(codeAction);
+                }
+            }
+        }
+        return actions;
+    }
+
+    private static CodeAction getFunctionImportCommand(LSDocumentIdentifier document, Diagnostic diagnostic,
+                                                       LSContext context) {
+        String diagnosticMessage = diagnostic.getMessage();
+        Position position = diagnostic.getRange().getStart();
+        int line = position.getLine();
+        int column = position.getCharacter();
+        String uri = context.get(CodeActionKeys.FILE_URI_KEY);
+        CommandArgument lineArg = new CommandArgument(CommandConstants.ARG_KEY_NODE_LINE, "" + line);
+        CommandArgument colArg = new CommandArgument(CommandConstants.ARG_KEY_NODE_COLUMN, "" + column);
+        CommandArgument uriArg = new CommandArgument(CommandConstants.ARG_KEY_DOC_URI, uri);
+        List<Diagnostic> diagnostics = new ArrayList<>();
+
+        List<Object> args = Arrays.asList(lineArg, colArg, uriArg);
+        Matcher matcher = CommandConstants.UNDEFINED_FUNCTION_PATTERN.matcher(diagnosticMessage);
+        String functionName = (matcher.find() && matcher.groupCount() > 0) ? matcher.group(1) + "(...)" : "";
+        WorkspaceDocumentManager docManager = context.get(CodeActionKeys.DOCUMENT_MANAGER_KEY);
+        try {
+            BLangInvocation node = CommandUtil.getFunctionInvocationNode(line, column, document.getURIString(),
+                                                                         docManager,
+                                                                         context);
+            if (node != null && node.pkgAlias.value.isEmpty()) {
+                boolean isWithinProject = (node.expr == null);
+                if (node.expr != null) {
+                    BLangPackage bLangPackage = context.get(DocumentServiceKeys.CURRENT_BLANG_PACKAGE_CONTEXT_KEY);
+                    List<String> currentModules = document.getProjectModules();
+                    PackageID nodePkgId = node.expr.type.tsymbol.pkgID;
+                    isWithinProject = bLangPackage.packageID.orgName.equals(nodePkgId.orgName) &&
+                            currentModules.contains(nodePkgId.name.value);
+                }
+                if (isWithinProject) {
+                    String commandTitle = CommandConstants.CREATE_FUNCTION_TITLE + functionName;
+                    CodeAction action = new CodeAction(commandTitle);
+                    action.setKind(CodeActionKind.QuickFix);
+                    action.setCommand(new Command(commandTitle, CreateFunctionExecutor.COMMAND, args));
+                    action.setDiagnostics(diagnostics);
+                    return action;
+                }
+            }
+        } catch (CompilationFailedException e) {
+            // ignore
+        }
+        return null;
+    }
+}
