+++ language-server/modules/langserver-core/src/main/java/org/ballerinalang/langserver/codeaction/providers/TaintedParamCodeAction.java
@@ -20,12 +20,25 @@ import org.ballerinalang.langserver.command.CommandUtil;
 import org.ballerinalang.langserver.common.constants.CommandConstants;
 import org.ballerinalang.langserver.commons.LSContext;
 import org.ballerinalang.langserver.commons.codeaction.CodeActionNodeType;
+import org.ballerinalang.langserver.commons.workspace.WorkspaceDocumentException;
+import org.ballerinalang.langserver.commons.workspace.WorkspaceDocumentManager;
+import org.ballerinalang.langserver.compiler.DocumentServiceKeys;
 import org.eclipse.lsp4j.CodeAction;
+import org.eclipse.lsp4j.CodeActionKind;
 import org.eclipse.lsp4j.Diagnostic;
+import org.eclipse.lsp4j.Range;
+import org.eclipse.lsp4j.TextDocumentEdit;
+import org.eclipse.lsp4j.TextEdit;
+import org.eclipse.lsp4j.VersionedTextDocumentIdentifier;
+import org.eclipse.lsp4j.WorkspaceEdit;
+import org.eclipse.lsp4j.jsonrpc.messages.Either;
 
+import java.io.IOException;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 import java.util.Locale;
+import java.util.regex.Matcher;
 
 /**
  * Code Action provider for tainted parameters.
@@ -42,15 +55,48 @@ public class TaintedParamCodeAction extends AbstractCodeActionProvider {
     public List<CodeAction> getCodeActions(CodeActionNodeType nodeType, LSContext lsContext,
                                            List<Diagnostic> diagnostics) {
         List<CodeAction> actions = new ArrayList<>();
-
         for (Diagnostic diagnostic : diagnostics) {
             if (diagnostic.getMessage().toLowerCase(Locale.ROOT).contains(CommandConstants.TAINTED_PARAM_PASSED)) {
-                CodeAction codeAction = CommandUtil.getTaintedParamCommand(diagnostic, lsContext);
+                CodeAction codeAction = getTaintedParamCommand(diagnostic, lsContext);
                 if (codeAction != null) {
                     actions.add(codeAction);
                 }
             }
         }
         return actions;
+    }
+
+    private static CodeAction getTaintedParamCommand(Diagnostic diagnostic, LSContext context) {
+        String diagnosticMessage = diagnostic.getMessage();
+        List<Diagnostic> diagnostics = new ArrayList<>();
+        String uri = context.get(DocumentServiceKeys.FILE_URI_KEY);
+
+        try {
+            Matcher matcher = CommandConstants.TAINTED_PARAM_PATTERN.matcher(diagnosticMessage);
+            if (matcher.find() && matcher.groupCount() > 0) {
+                String param = matcher.group(1);
+                String commandTitle = String.format(CommandConstants.MARK_UNTAINTED_TITLE, param);
+                CodeAction action = new CodeAction(commandTitle);
+                action.setKind(CodeActionKind.QuickFix);
+                action.setDiagnostics(diagnostics);
+                // Extract specific content range
+                Range range = diagnostic.getRange();
+                WorkspaceDocumentManager documentManager = context.get(DocumentServiceKeys.DOC_MANAGER_KEY);
+                String content = CommandUtil.getContentOfRange(documentManager, uri, range);
+                // Add `untaint` keyword
+                matcher = CommandConstants.NO_CONCAT_PATTERN.matcher(content);
+                String editText = matcher.find() ? "<@untained>  " + content : "<@untained> (" + content + ")";
+                // Create text-edit
+                List<TextEdit> edits = new ArrayList<>();
+                edits.add(new TextEdit(range, editText));
+                VersionedTextDocumentIdentifier identifier = new VersionedTextDocumentIdentifier(uri, null);
+                action.setEdit(new WorkspaceEdit(Collections.singletonList(
+                        Either.forLeft(new TextDocumentEdit(identifier, edits)))));
+                return action;
+            }
+        } catch (WorkspaceDocumentException | IOException e) {
+            //do nothing
+        }
+        return null;
     }
 }
