+++ language-server/modules/langserver-core/src/main/java/org/ballerinalang/langserver/command/CommandUtil.java
@@ -15,36 +15,20 @@
  */
 package org.ballerinalang.langserver.command;
 
-import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.tuple.ImmutablePair;
 import org.apache.commons.lang3.tuple.Pair;
-import org.ballerinalang.jvm.util.BLangConstants;
 import org.ballerinalang.langserver.client.ExtendedLanguageClient;
-import org.ballerinalang.langserver.command.executors.AddAllDocumentationExecutor;
-import org.ballerinalang.langserver.command.executors.AddDocumentationExecutor;
-import org.ballerinalang.langserver.command.executors.ChangeAbstractTypeObjExecutor;
-import org.ballerinalang.langserver.command.executors.CreateFunctionExecutor;
-import org.ballerinalang.langserver.command.executors.CreateTestExecutor;
-import org.ballerinalang.langserver.command.executors.ImportModuleExecutor;
-import org.ballerinalang.langserver.command.executors.PullModuleExecutor;
-import org.ballerinalang.langserver.common.CommonKeys;
-import org.ballerinalang.langserver.common.constants.CommandConstants;
 import org.ballerinalang.langserver.common.constants.NodeContextKeys;
 import org.ballerinalang.langserver.common.position.PositionTreeVisitor;
 import org.ballerinalang.langserver.common.utils.CommonUtil;
-import org.ballerinalang.langserver.common.utils.FunctionGenerator;
 import org.ballerinalang.langserver.commons.LSContext;
-import org.ballerinalang.langserver.commons.codeaction.CodeActionKeys;
 import org.ballerinalang.langserver.commons.workspace.LSDocumentIdentifier;
 import org.ballerinalang.langserver.commons.workspace.WorkspaceDocumentException;
 import org.ballerinalang.langserver.commons.workspace.WorkspaceDocumentManager;
 import org.ballerinalang.langserver.compiler.DocumentServiceKeys;
-import org.ballerinalang.langserver.compiler.LSCompilerUtil;
 import org.ballerinalang.langserver.compiler.LSModuleCompiler;
-import org.ballerinalang.langserver.compiler.LSPackageLoader;
 import org.ballerinalang.langserver.compiler.common.LSCustomErrorStrategy;
 import org.ballerinalang.langserver.compiler.common.LSDocumentIdentifierImpl;
-import org.ballerinalang.langserver.compiler.common.modal.BallerinaPackage;
 import org.ballerinalang.langserver.compiler.exception.CompilationFailedException;
 import org.ballerinalang.langserver.diagnostic.DiagnosticsHelper;
 import org.ballerinalang.langserver.util.references.SymbolReferencesModel;
@@ -54,10 +38,6 @@ import org.ballerinalang.model.tree.TopLevelNode;
 import org.ballerinalang.model.tree.expressions.IndexBasedAccessNode;
 import org.ballerinalang.model.types.TypeKind;
 import org.eclipse.lsp4j.ApplyWorkspaceEditParams;
-import org.eclipse.lsp4j.CodeAction;
-import org.eclipse.lsp4j.CodeActionKind;
-import org.eclipse.lsp4j.Command;
-import org.eclipse.lsp4j.Diagnostic;
 import org.eclipse.lsp4j.MessageParams;
 import org.eclipse.lsp4j.MessageType;
 import org.eclipse.lsp4j.Position;
@@ -71,56 +51,21 @@ import org.eclipse.lsp4j.VersionedTextDocumentIdentifier;
 import org.eclipse.lsp4j.WorkspaceEdit;
 import org.eclipse.lsp4j.jsonrpc.messages.Either;
 import org.eclipse.lsp4j.services.LanguageClient;
-import org.wso2.ballerinalang.compiler.semantics.model.symbols.BInvokableSymbol;
-import org.wso2.ballerinalang.compiler.semantics.model.symbols.BObjectTypeSymbol;
-import org.wso2.ballerinalang.compiler.semantics.model.symbols.BSymbol;
-import org.wso2.ballerinalang.compiler.semantics.model.types.BErrorType;
-import org.wso2.ballerinalang.compiler.semantics.model.types.BNilType;
-import org.wso2.ballerinalang.compiler.semantics.model.types.BType;
-import org.wso2.ballerinalang.compiler.semantics.model.types.BUnionType;
-import org.wso2.ballerinalang.compiler.tree.BLangCompilationUnit;
-import org.wso2.ballerinalang.compiler.tree.BLangFunction;
-import org.wso2.ballerinalang.compiler.tree.BLangImportPackage;
 import org.wso2.ballerinalang.compiler.tree.BLangNode;
 import org.wso2.ballerinalang.compiler.tree.BLangPackage;
-import org.wso2.ballerinalang.compiler.tree.BLangService;
-import org.wso2.ballerinalang.compiler.tree.BLangTypeDefinition;
 import org.wso2.ballerinalang.compiler.tree.expressions.BLangInvocation;
-import org.wso2.ballerinalang.compiler.tree.statements.BLangBlockStmt;
-import org.wso2.ballerinalang.compiler.tree.statements.BLangReturn;
-import org.wso2.ballerinalang.compiler.tree.statements.BLangStatement;
 import org.wso2.ballerinalang.compiler.tree.types.BLangObjectTypeNode;
-import org.wso2.ballerinalang.compiler.tree.types.BLangValueType;
-import org.wso2.ballerinalang.compiler.util.CompilerContext;
-import org.wso2.ballerinalang.compiler.util.diagnotic.DiagnosticPos;
-import org.wso2.ballerinalang.util.Flags;
 
 import java.io.BufferedReader;
-import java.io.File;
 import java.io.IOException;
 import java.io.StringReader;
-import java.lang.reflect.Field;
 import java.nio.file.Path;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collections;
-import java.util.Iterator;
 import java.util.List;
-import java.util.Locale;
 import java.util.Optional;
-import java.util.Set;
-import java.util.function.BiConsumer;
-import java.util.function.Function;
-import java.util.regex.Matcher;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
-import java.util.stream.Stream;
 
-import static org.ballerinalang.langserver.common.utils.CommonUtil.LINE_SEPARATOR;
-import static org.ballerinalang.langserver.common.utils.CommonUtil.createVariableDeclaration;
 import static org.ballerinalang.langserver.compiler.LSClientLogger.logError;
 import static org.ballerinalang.langserver.compiler.LSCompilerUtil.getUntitledFilePath;
-import static org.ballerinalang.langserver.util.references.ReferencesUtil.getReferenceAtCursor;
 
 /**
  * Utilities for the command related operations.
@@ -128,62 +73,6 @@ import static org.ballerinalang.langserver.util.references.ReferencesUtil.getRef
 public class CommandUtil {
 
     private CommandUtil() {
-    }
-
-    /**
-     * Get the command for generate test class.
-     *
-     * @param topLevelNodeType top level node
-     * @param docUri           Document URI
-     * @param context          {@link LSContext}
-     * @return {@link Command} Test Generation command
-     * @throws CompilationFailedException thrown when compilation failed
-     */
-    public static List<CodeAction> getTestGenerationCommand(String topLevelNodeType, String docUri,
-                                                            LSContext context)
-            throws CompilationFailedException {
-        List<CodeAction> actions = new ArrayList<>();
-        List<Object> args = new ArrayList<>();
-        args.add(new CommandArgument(CommandConstants.ARG_KEY_DOC_URI, docUri));
-        Position position = context.get(DocumentServiceKeys.POSITION_KEY).getPosition();
-        args.add(new CommandArgument(CommandConstants.ARG_KEY_NODE_LINE, "" + position.getLine()));
-        args.add(new CommandArgument(CommandConstants.ARG_KEY_NODE_COLUMN, "" + position.getCharacter()));
-
-        boolean isService = CommonKeys.SERVICE_KEYWORD_KEY.equals(topLevelNodeType);
-        boolean isFunction = CommonKeys.FUNCTION_KEYWORD_KEY.equals(topLevelNodeType);
-        WorkspaceDocumentManager documentManager = context.get(DocumentServiceKeys.DOC_MANAGER_KEY);
-        if ((isService || isFunction) && !isTopLevelNode(docUri, documentManager, context, position)) {
-            return actions;
-        }
-
-        if (isService) {
-            CodeAction action = new CodeAction(CommandConstants.CREATE_TEST_SERVICE_TITLE);
-            action.setCommand(new Command(CommandConstants.CREATE_TEST_SERVICE_TITLE,
-                                          CreateTestExecutor.COMMAND, args));
-            actions.add(action);
-        } else if (isFunction) {
-            CodeAction action = new CodeAction(CommandConstants.CREATE_TEST_FUNC_TITLE);
-            action.setCommand(new Command(CommandConstants.CREATE_TEST_FUNC_TITLE,
-                                          CreateTestExecutor.COMMAND, args));
-            actions.add(action);
-        }
-        return actions;
-    }
-
-    private static boolean isTopLevelNode(String uri, WorkspaceDocumentManager docManager, LSContext context,
-                                          Position position)
-            throws CompilationFailedException {
-        Pair<BLangNode, Object> bLangNode = getBLangNode(position.getLine(), position.getCharacter(), uri, docManager,
-                                                         context);
-
-        // Only supported for 'public' functions
-        if (bLangNode.getLeft() instanceof BLangFunction &&
-                !((BLangFunction) bLangNode.getLeft()).getFlags().contains(Flag.PUBLIC)) {
-            return false;
-        }
-
-        // Only supported for top-level nodes
-        return (bLangNode.getLeft() != null && bLangNode.getLeft().parent instanceof BLangPackage);
     }
 
     public static Pair<BLangNode, Object> getBLangNode(int line, int column, String uri,
@@ -305,68 +194,6 @@ public class CommandUtil {
         }
     }
 
-    public static CodeAction getDocGenerationCommand(String nodeType, String docUri, int line) {
-        CommandArgument nodeTypeArg = new CommandArgument(CommandConstants.ARG_KEY_NODE_TYPE, nodeType);
-        CommandArgument docUriArg = new CommandArgument(CommandConstants.ARG_KEY_DOC_URI, docUri);
-        CommandArgument lineStart = new CommandArgument(CommandConstants.ARG_KEY_NODE_LINE, String.valueOf(line));
-        List<Object> args = new ArrayList<>(Arrays.asList(nodeTypeArg, docUriArg, lineStart));
-        CodeAction action = new CodeAction(CommandConstants.ADD_DOCUMENTATION_TITLE);
-        action.setCommand(new Command(CommandConstants.ADD_DOCUMENTATION_TITLE,
-                                      AddDocumentationExecutor.COMMAND, args));
-        return action;
-    }
-
-    public static CodeAction getAllDocGenerationCommand(String docUri) {
-        CommandArgument docUriArg = new CommandArgument(CommandConstants.ARG_KEY_DOC_URI, docUri);
-        List<Object> args = new ArrayList<>(Collections.singletonList(docUriArg));
-        CodeAction action = new CodeAction(CommandConstants.ADD_ALL_DOC_TITLE);
-        action.setCommand(new Command(CommandConstants.ADD_ALL_DOC_TITLE, AddAllDocumentationExecutor.COMMAND, args));
-        return action;
-    }
-
-    public static CodeAction getFunctionImportCommand(LSDocumentIdentifier document, Diagnostic diagnostic,
-                                                      LSContext context) {
-        String diagnosticMessage = diagnostic.getMessage();
-        Position position = diagnostic.getRange().getStart();
-        int line = position.getLine();
-        int column = position.getCharacter();
-        String uri = context.get(CodeActionKeys.FILE_URI_KEY);
-        CommandArgument lineArg = new CommandArgument(CommandConstants.ARG_KEY_NODE_LINE, "" + line);
-        CommandArgument colArg = new CommandArgument(CommandConstants.ARG_KEY_NODE_COLUMN, "" + column);
-        CommandArgument uriArg = new CommandArgument(CommandConstants.ARG_KEY_DOC_URI, uri);
-        List<Diagnostic> diagnostics = new ArrayList<>();
-
-        List<Object> args = Arrays.asList(lineArg, colArg, uriArg);
-        Matcher matcher = CommandConstants.UNDEFINED_FUNCTION_PATTERN.matcher(diagnosticMessage);
-        String functionName = (matcher.find() && matcher.groupCount() > 0) ? matcher.group(1) + "(...)" : "";
-        WorkspaceDocumentManager docManager = context.get(CodeActionKeys.DOCUMENT_MANAGER_KEY);
-        try {
-            BLangInvocation node = getFunctionInvocationNode(line, column, document.getURIString(), docManager,
-                                                             context);
-            if (node != null && node.pkgAlias.value.isEmpty()) {
-                boolean isWithinProject = (node.expr == null);
-                if (node.expr != null) {
-                    BLangPackage bLangPackage = context.get(DocumentServiceKeys.CURRENT_BLANG_PACKAGE_CONTEXT_KEY);
-                    List<String> currentModules = document.getProjectModules();
-                    PackageID nodePkgId = node.expr.type.tsymbol.pkgID;
-                    isWithinProject = bLangPackage.packageID.orgName.equals(nodePkgId.orgName) &&
-                            currentModules.contains(nodePkgId.name.value);
-                }
-                if (isWithinProject) {
-                    String commandTitle = CommandConstants.CREATE_FUNCTION_TITLE + functionName;
-                    CodeAction action = new CodeAction(commandTitle);
-                    action.setKind(CodeActionKind.QuickFix);
-                    action.setCommand(new Command(commandTitle, CreateFunctionExecutor.COMMAND, args));
-                    action.setDiagnostics(diagnostics);
-                    return action;
-                }
-            }
-        } catch (CompilationFailedException e) {
-            // ignore
-        }
-        return null;
-    }
-
     public static BLangInvocation getFunctionInvocationNode(int line, int column, String uri,
                                                             WorkspaceDocumentManager documentManager, LSContext context)
             throws CompilationFailedException {
@@ -387,408 +214,17 @@ public class CommandUtil {
         }
     }
 
-    public static List<CodeAction> getModuleImportCommand(Diagnostic diagnostic,
-                                                          LSContext context) {
-        String diagnosticMessage = diagnostic.getMessage();
-        List<CodeAction> actions = new ArrayList<>();
-        String uri = context.get(DocumentServiceKeys.FILE_URI_KEY);
-        CommandArgument uriArg = new CommandArgument(CommandConstants.ARG_KEY_DOC_URI, uri);
-        List<Diagnostic> diagnostics = new ArrayList<>();
-
-        String packageAlias = diagnosticMessage.substring(diagnosticMessage.indexOf("'") + 1,
-                                                          diagnosticMessage.lastIndexOf("'"));
-        LSDocumentIdentifier sourceDocument = new LSDocumentIdentifierImpl(uri);
-        String sourceRoot = LSCompilerUtil.getProjectRoot(sourceDocument.getPath());
-        sourceDocument.setProjectRootRoot(sourceRoot);
-        List<BallerinaPackage> packagesList = new ArrayList<>();
-        Stream.of(LSPackageLoader.getSdkPackages(), LSPackageLoader.getHomeRepoPackages())
-                .forEach(packagesList::addAll);
-        packagesList.stream()
-                .filter(pkgEntry -> {
-                    String fullPkgName = pkgEntry.getFullPackageNameAlias();
-                    return fullPkgName.endsWith("." + packageAlias) || fullPkgName.endsWith("/" + packageAlias);
-                })
-                .forEach(pkgEntry -> {
-                    String commandTitle = CommandConstants.IMPORT_MODULE_TITLE + " "
-                            + pkgEntry.getFullPackageNameAlias();
-                    CommandArgument pkgArgument = new CommandArgument(CommandConstants.ARG_KEY_MODULE_NAME,
-                                                                      pkgEntry.getFullPackageNameAlias());
-                    CodeAction action = new CodeAction(commandTitle);
-                    action.setKind(CodeActionKind.QuickFix);
-                    action.setCommand(new Command(commandTitle, ImportModuleExecutor.COMMAND,
-                                                  new ArrayList<>(Arrays.asList(pkgArgument, uriArg))));
-                    action.setDiagnostics(diagnostics);
-                    actions.add(action);
-                });
-        return actions;
-    }
-
-    public static CodeAction getUnresolvedModulesCommand(Diagnostic diagnostic, LSContext context) {
-        String diagnosticMessage = diagnostic.getMessage();
-        String uri = context.get(DocumentServiceKeys.FILE_URI_KEY);
-        CommandArgument uriArg = new CommandArgument(CommandConstants.ARG_KEY_DOC_URI, uri);
-
-        Matcher matcher = CommandConstants.UNRESOLVED_MODULE_PATTERN.matcher(
-                diagnosticMessage.toLowerCase(Locale.ROOT)
-        );
-        List<Diagnostic> diagnostics = new ArrayList<>();
-        if (matcher.find() && matcher.groupCount() > 0) {
-            List<Object> args = new ArrayList<>();
-            String pkgName = matcher.group(1);
-            args.add(new CommandArgument(CommandConstants.ARG_KEY_MODULE_NAME, pkgName));
-            args.add(uriArg);
-            String commandTitle = CommandConstants.PULL_MOD_TITLE;
-            CodeAction action = new CodeAction(commandTitle);
-            action.setKind(CodeActionKind.QuickFix);
-            action.setCommand(new Command(commandTitle, PullModuleExecutor.COMMAND, args));
-            action.setDiagnostics(diagnostics);
-            return action;
-        }
-        return null;
-    }
-
-    public static CodeAction getChangeAbstractTypeCommand(Diagnostic diagnostic, LSContext context) {
-        String diagnosticMessage = diagnostic.getMessage();
-        Position position = diagnostic.getRange().getStart();
-        int line = position.getLine();
-        int column = position.getCharacter();
-        String uri = context.get(CodeActionKeys.FILE_URI_KEY);
-        CommandArgument lineArg = new CommandArgument(CommandConstants.ARG_KEY_NODE_LINE, "" + line);
-        CommandArgument colArg = new CommandArgument(CommandConstants.ARG_KEY_NODE_COLUMN, "" + column);
-        CommandArgument uriArg = new CommandArgument(CommandConstants.ARG_KEY_DOC_URI, uri);
-        List<Diagnostic> diagnostics = new ArrayList<>();
-
-        Matcher matcher = CommandConstants.FUNC_IN_ABSTRACT_OBJ_PATTERN.matcher(diagnosticMessage);
-        if (matcher.find() && matcher.groupCount() > 1) {
-            String objectName = matcher.group(2);
-            int colonIndex = objectName.lastIndexOf(":");
-            String simpleObjName = (colonIndex > -1) ? objectName.substring(colonIndex + 1) : objectName;
-            List<Object> args = Arrays.asList(lineArg, colArg, uriArg);
-            String commandTitle = String.format(CommandConstants.MAKE_OBJ_NON_ABSTRACT_TITLE, simpleObjName);
-
-            CodeAction action = new CodeAction(commandTitle);
-            action.setKind(CodeActionKind.QuickFix);
-            action.setCommand(new Command(commandTitle, ChangeAbstractTypeObjExecutor.COMMAND, args));
-            action.setDiagnostics(diagnostics);
-            return action;
-        }
-        return null;
-    }
-
-    public static CodeAction getNoImplementationFoundCommand(Diagnostic diagnostic, LSContext context) {
-        String diagnosticMessage = diagnostic.getMessage();
-        Position position = diagnostic.getRange().getStart();
-        int line = position.getLine();
-        int column = position.getCharacter();
-        String uri = context.get(CodeActionKeys.FILE_URI_KEY);
-        CommandArgument lineArg = new CommandArgument(CommandConstants.ARG_KEY_NODE_LINE, "" + line);
-        CommandArgument colArg = new CommandArgument(CommandConstants.ARG_KEY_NODE_COLUMN, "" + column);
-        CommandArgument uriArg = new CommandArgument(CommandConstants.ARG_KEY_DOC_URI, uri);
-        List<Diagnostic> diagnostics = new ArrayList<>();
-
-        Matcher matcher = CommandConstants.NO_IMPL_FOUND_FOR_FUNCTION_PATTERN.matcher(diagnosticMessage);
-        if (matcher.find() && matcher.groupCount() > 1) {
-            String objectName = matcher.group(2);
-            int colonIndex = objectName.lastIndexOf(":");
-            String simpleObjName = (colonIndex > -1) ? objectName.substring(colonIndex + 1) : objectName;
-            List<Object> args = Arrays.asList(lineArg, colArg, uriArg);
-            String commandTitle = String.format(CommandConstants.MAKE_OBJ_ABSTRACT_TITLE, simpleObjName);
-
-            CodeAction action = new CodeAction(commandTitle);
-            action.setKind(CodeActionKind.QuickFix);
-            action.setCommand(new Command(commandTitle, ChangeAbstractTypeObjExecutor.COMMAND, args));
-            action.setDiagnostics(diagnostics);
-            return action;
-        }
-        return null;
-    }
-
-    public static CodeAction getUnresolvedPackageCommand(Diagnostic diagnostic, LSContext context) {
-        String diagnosticMessage = diagnostic.getMessage();
-        String uri = context.get(CodeActionKeys.FILE_URI_KEY);
-        CommandArgument uriArg = new CommandArgument(CommandConstants.ARG_KEY_DOC_URI, uri);
-        List<Diagnostic> diagnostics = new ArrayList<>();
-
-        Matcher matcher = CommandConstants.UNRESOLVED_MODULE_PATTERN.matcher(
-                diagnosticMessage.toLowerCase(Locale.ROOT)
-        );
-        if (matcher.find() && matcher.groupCount() > 0) {
-            List<Object> args = new ArrayList<>();
-            String pkgName = matcher.group(1);
-            args.add(new CommandArgument(CommandConstants.ARG_KEY_MODULE_NAME, pkgName));
-            args.add(uriArg);
-            String commandTitle = CommandConstants.PULL_MOD_TITLE;
-            CodeAction action = new CodeAction(commandTitle);
-            action.setKind(CodeActionKind.QuickFix);
-            action.setCommand(new Command(commandTitle, PullModuleExecutor.COMMAND, args));
-            action.setDiagnostics(diagnostics);
-            return action;
-        }
-        return null;
-    }
-
-    public static CodeAction getIncompatibleTypesCommand(LSDocumentIdentifier document, Diagnostic diagnostic,
-                                                         LSContext context) {
-        String diagnosticMessage = diagnostic.getMessage();
-        Position position = diagnostic.getRange().getStart();
-        int line = position.getLine();
-        int column = position.getCharacter();
-        String uri = context.get(CodeActionKeys.FILE_URI_KEY);
-        List<Diagnostic> diagnostics = new ArrayList<>();
-
-        Matcher matcher = CommandConstants.INCOMPATIBLE_TYPE_PATTERN.matcher(diagnosticMessage);
-        if (matcher.find() && matcher.groupCount() > 1) {
-            String foundType = matcher.group(2);
-            WorkspaceDocumentManager documentManager = context.get(DocumentServiceKeys.DOC_MANAGER_KEY);
-            try {
-                BLangFunction func = CommandUtil.getFunctionNode(line, column, document, documentManager, context);
-                if (func != null && !BLangConstants.MAIN_FUNCTION_NAME.equals(func.name.value)) {
-                    BLangStatement statement = CommandUtil.getStatementByLocation(func.getBody().getStatements(),
-                                                                                  line + 1, column + 1);
-                    if (statement instanceof BLangReturn) {
-                        // Process full-qualified BType name  eg. ballerina/http:Client and if required; add
-                        // auto-import
-                        matcher = CommandConstants.FQ_TYPE_PATTERN.matcher(foundType);
-                        List<TextEdit> edits = new ArrayList<>();
-                        String editText = extractTypeName(context, matcher, foundType, edits);
-
-                        // Process function node
-                        Position start;
-                        Position end;
-                        if (func.returnTypeNode instanceof BLangValueType
-                                && TypeKind.NIL.equals(((BLangValueType) func.returnTypeNode).getTypeKind())
-                                && func.returnTypeNode.getWS() == null) {
-                            // eg. function test() {...}
-                            start = new Position(func.returnTypeNode.pos.sLine - 1,
-                                                 func.returnTypeNode.pos.eCol - 1);
-                            end = new Position(func.returnTypeNode.pos.eLine - 1, func.returnTypeNode.pos.eCol - 1);
-                            editText = " returns (" + editText + ")";
-                        } else {
-                            // eg. function test() returns () {...}
-                            start = new Position(func.returnTypeNode.pos.sLine - 1,
-                                                 func.returnTypeNode.pos.sCol - 1);
-                            end = new Position(func.returnTypeNode.pos.eLine - 1, func.returnTypeNode.pos.eCol - 1);
-                        }
-                        edits.add(new TextEdit(new Range(start, end), editText));
-
-                        // Add code action
-                        String commandTitle = CommandConstants.CHANGE_RETURN_TYPE_TITLE + foundType + "'";
-                        CodeAction action = new CodeAction(commandTitle);
-                        action.setKind(CodeActionKind.QuickFix);
-                        action.setDiagnostics(diagnostics);
-                        action.setEdit(new WorkspaceEdit(Collections.singletonList(
-                                Either.forLeft(new TextDocumentEdit(new VersionedTextDocumentIdentifier(uri, null),
-                                                                    edits))))
-                        );
-                        return action;
-                    }
-                }
-            } catch (CompilationFailedException e) {
-                // ignore
-            }
-        }
-        return null;
-    }
-
-    private static BLangFunction getFunctionNode(int line, int column, LSDocumentIdentifier document,
-                                                 WorkspaceDocumentManager docManager, LSContext context)
-            throws CompilationFailedException {
-        String uri = document.getURIString();
-        Position position = new Position();
-        position.setLine(line);
-        position.setCharacter(column + 1);
-        context.put(DocumentServiceKeys.FILE_URI_KEY, uri);
-        TextDocumentIdentifier identifier = new TextDocumentIdentifier(uri);
-        context.put(DocumentServiceKeys.POSITION_KEY, new TextDocumentPositionParams(identifier, position));
-        LSModuleCompiler.getBLangPackages(context, docManager, LSCustomErrorStrategy.class, true, false, false);
-
-        // Get the current package.
-        BLangPackage currentPackage = context.get(DocumentServiceKeys.CURRENT_BLANG_PACKAGE_CONTEXT_KEY);
-
-        // If package is testable package process as tests
-        // else process normally
-        String relativeFilePath = context.get(DocumentServiceKeys.RELATIVE_FILE_PATH_KEY);
-        BLangCompilationUnit compilationUnit;
-        if (relativeFilePath.startsWith("tests" + File.separator)) {
-            compilationUnit = currentPackage.getTestablePkg().getCompilationUnits().stream().
-                    filter(compUnit -> (relativeFilePath).equals(compUnit.getName()))
-                    .findFirst().orElse(null);
-        } else {
-            compilationUnit = currentPackage.getCompilationUnits().stream().
-                    filter(compUnit -> relativeFilePath.equals(compUnit.getName())).findFirst().orElse(null);
-        }
-        if (compilationUnit == null) {
-            return null;
-        }
-        Iterator<TopLevelNode> nodeIterator = compilationUnit.getTopLevelNodes().iterator();
-        BLangFunction result = null;
-        TopLevelNode next = (nodeIterator.hasNext()) ? nodeIterator.next() : null;
-        Function<org.ballerinalang.util.diagnostic.Diagnostic.DiagnosticPosition, Boolean> isWithinPosition =
-                diagnosticPosition -> {
-                    int sLine = diagnosticPosition.getStartLine();
-                    int eLine = diagnosticPosition.getEndLine();
-                    int sCol = diagnosticPosition.getStartColumn();
-                    int eCol = diagnosticPosition.getEndColumn();
-                    return ((line > sLine || (line == sLine && column >= sCol)) &&
-                            (line < eLine || (line == eLine && column <= eCol)));
-                };
-        while (next != null) {
-            if (isWithinPosition.apply(next.getPosition())) {
-                if (next instanceof BLangFunction) {
-                    result = (BLangFunction) next;
-                    break;
-                } else if (next instanceof BLangTypeDefinition) {
-                    BLangTypeDefinition typeDefinition = (BLangTypeDefinition) next;
-                    if (typeDefinition.typeNode instanceof BLangObjectTypeNode) {
-                        BLangObjectTypeNode typeNode = (BLangObjectTypeNode) typeDefinition.typeNode;
-                        for (BLangFunction function : typeNode.functions) {
-                            if (isWithinPosition.apply(function.getPosition())) {
-                                result = function;
-                                break;
-                            }
-                        }
-                    }
-                } else if (next instanceof BLangService) {
-                    BLangService bLangService = (BLangService) next;
-                    for (BLangFunction function : bLangService.resourceFunctions) {
-                        if (isWithinPosition.apply(function.getPosition())) {
-                            result = function;
-                            break;
-                        }
-                    }
-                }
-                break;
-            }
-            next = (nodeIterator.hasNext()) ? nodeIterator.next() : null;
-        }
-        return result;
-    }
-
     /**
-     * Find statement by location.
+     * Get content of range of a given file uri.
      *
-     * @param statements list of statements
-     * @param line       line
-     * @param column     column
-     * @return {@link BLangStatement} if found, NULL otherwise
+     * @param documentManager   document manager
+     * @param uri   file uri
+     * @param range content range
+     * @return content of range
+     * @throws WorkspaceDocumentException   when file not found
+     * @throws IOException  when I/O error occurs
      */
-    private static BLangStatement getStatementByLocation(List<BLangStatement> statements, int line, int column) {
-        BLangStatement node = null;
-        for (BLangStatement statement : statements) {
-            BLangStatement rv;
-            if ((rv = getStatementByLocation(statement, line, column)) != null) {
-                return rv;
-            }
-        }
-        return node;
-    }
-
-    /**
-     * Find statements by location.
-     *
-     * @param node   lookup {@link BLangNode}
-     * @param line   line
-     * @param column column
-     * @return {@link BLangStatement} if found, NULL otherwise
-     */
-    private static BLangStatement getStatementByLocation(BLangNode node, int line, int column) {
-        try {
-            if (checkNodeWithin(node, line, column) && node instanceof BLangStatement) {
-                return (BLangStatement) node;
-            }
-            for (Field field : node.getClass().getDeclaredFields()) {
-                Object obj = field.get(node);
-                if (obj instanceof BLangBlockStmt) {
-                    // Found a block-statement field, drilling down further
-                    BLangStatement rv;
-                    if ((rv = getStatementByLocation(((BLangBlockStmt) obj).getStatements(), line, column)) != null) {
-                        return rv;
-                    }
-                } else if (obj instanceof BLangStatement) {
-                    if (checkNodeWithin((BLangStatement) obj, line, column)) {
-                        return (BLangStatement) obj;
-                    }
-                    // Found a statement field, drilling down further
-                    BLangStatement rv;
-                    if ((rv = getStatementByLocation((BLangStatement) obj, line, column)) != null) {
-                        return rv;
-                    }
-                }
-            }
-        } catch (IllegalArgumentException | IllegalAccessException e) {
-            return null;
-        }
-        return null;
-    }
-
-    private static boolean checkNodeWithin(BLangNode node, int line, int column) {
-        int sLine = node.getPosition().getStartLine();
-        int eLine = node.getPosition().getEndLine();
-        int sCol = node.getPosition().getStartColumn();
-        int eCol = node.getPosition().getEndColumn();
-        return (line > sLine || (line == sLine && column >= sCol)) &&
-                (line < eLine || (line == eLine && column <= eCol));
-    }
-
-    private static String extractTypeName(LSContext context, Matcher matcher, String foundType, List<TextEdit> edits) {
-        if (matcher.find() && matcher.groupCount() > 2) {
-            String orgName = matcher.group(1);
-            String alias = matcher.group(2);
-            String typeName = matcher.group(3);
-            String pkgId = orgName + "/" + alias;
-            PackageID currentPkgId = context.get(
-                    DocumentServiceKeys.CURRENT_BLANG_PACKAGE_CONTEXT_KEY).packageID;
-            if (pkgId.equals(currentPkgId.toString())) {
-                foundType = typeName;
-            } else {
-                List<BLangImportPackage> currentDocImports = CommonUtil.getCurrentFileImports(context);
-                boolean pkgAlreadyImported = currentDocImports.stream()
-                        .anyMatch(importPkg -> importPkg.orgName.value.equals(orgName)
-                                && importPkg.alias.value.equals(alias));
-                if (!pkgAlreadyImported) {
-                    edits.addAll(CommonUtil.getAutoImportTextEdits(orgName, alias, context));
-                }
-                foundType = alias + CommonKeys.PKG_DELIMITER_KEYWORD + typeName;
-            }
-        }
-        return foundType;
-    }
-
-    public static CodeAction getTaintedParamCommand(Diagnostic diagnostic, LSContext context) {
-        String diagnosticMessage = diagnostic.getMessage();
-        List<Diagnostic> diagnostics = new ArrayList<>();
-        String uri = context.get(DocumentServiceKeys.FILE_URI_KEY);
-
-        try {
-            Matcher matcher = CommandConstants.TAINTED_PARAM_PATTERN.matcher(diagnosticMessage);
-            if (matcher.find() && matcher.groupCount() > 0) {
-                String param = matcher.group(1);
-                String commandTitle = String.format(CommandConstants.MARK_UNTAINTED_TITLE, param);
-                CodeAction action = new CodeAction(commandTitle);
-                action.setKind(CodeActionKind.QuickFix);
-                action.setDiagnostics(diagnostics);
-                // Extract specific content range
-                Range range = diagnostic.getRange();
-                WorkspaceDocumentManager documentManager = context.get(DocumentServiceKeys.DOC_MANAGER_KEY);
-                String content = getContentOfRange(documentManager, uri, range);
-                // Add `untaint` keyword
-                matcher = CommandConstants.NO_CONCAT_PATTERN.matcher(content);
-                String editText = matcher.find() ? "<@untained>  " + content : "<@untained> (" + content + ")";
-                // Create text-edit
-                List<TextEdit> edits = new ArrayList<>();
-                edits.add(new TextEdit(range, editText));
-                VersionedTextDocumentIdentifier identifier = new VersionedTextDocumentIdentifier(uri, null);
-                action.setEdit(new WorkspaceEdit(Collections.singletonList(
-                        Either.forLeft(new TextDocumentEdit(identifier, edits)))));
-                return action;
-            }
-        } catch (WorkspaceDocumentException | IOException e) {
-            //do nothing
-        }
-        return null;
-    }
-
-    private static String getContentOfRange(WorkspaceDocumentManager documentManager, String uri, Range range)
+    public static String getContentOfRange(WorkspaceDocumentManager documentManager, String uri, Range range)
             throws WorkspaceDocumentException, IOException {
         Optional<Path> filePath = CommonUtil.getPathFromURI(uri);
         if (!filePath.isPresent()) {
@@ -830,312 +266,5 @@ public class CommandUtil {
             lineNum++;
         }
         return capture.toString();
-    }
-
-    public static List<CodeAction> getVariableAssignmentCommand(LSDocumentIdentifier document, Diagnostic diagnostic,
-                                                                LSContext context) {
-        List<CodeAction> actions = new ArrayList<>();
-        String uri = context.get(DocumentServiceKeys.FILE_URI_KEY);
-        String diagnosedContent = getDiagnosedContent(diagnostic, context, document);
-        List<Diagnostic> diagnostics = new ArrayList<>();
-        Position position = diagnostic.getRange().getStart();
-
-        try {
-            Position afterAliasPos = offsetInvocation(diagnosedContent, position);
-            SymbolReferencesModel.Reference refAtCursor = getReferenceAtCursor(context, document, afterAliasPos);
-            BSymbol symbolAtCursor = refAtCursor.getSymbol();
-
-            boolean isInvocation = symbolAtCursor instanceof BInvokableSymbol;
-            boolean isRemoteInvocation = (symbolAtCursor.flags & Flags.REMOTE) == Flags.REMOTE;
-
-            boolean hasDefaultInitFunction = false;
-            boolean hasCustomInitFunction = false;
-            if (refAtCursor.getbLangNode() instanceof BLangInvocation) {
-                hasDefaultInitFunction = symbolAtCursor instanceof BObjectTypeSymbol;
-                hasCustomInitFunction = symbolAtCursor instanceof BInvokableSymbol &&
-                        symbolAtCursor.name.value.endsWith("__init");
-            }
-            boolean isInitInvocation = hasDefaultInitFunction || hasCustomInitFunction;
-
-            String commandTitle = CommandConstants.CREATE_VARIABLE_TITLE;
-            CodeAction action = new CodeAction(commandTitle);
-            List<TextEdit> edits = getCreateVariableCodeActionEdits(context, uri, refAtCursor,
-                                                                    hasDefaultInitFunction, hasCustomInitFunction);
-            action.setKind(CodeActionKind.QuickFix);
-            action.setEdit(new WorkspaceEdit(Collections.singletonList(Either.forLeft(
-                    new TextDocumentEdit(new VersionedTextDocumentIdentifier(uri, null), edits)))));
-            action.setDiagnostics(diagnostics);
-            actions.add(action);
-
-            if (isInvocation || isInitInvocation) {
-                BType returnType;
-                if (hasDefaultInitFunction) {
-                    returnType = symbolAtCursor.type;
-                } else if (hasCustomInitFunction) {
-                    returnType = symbolAtCursor.owner.type;
-                } else {
-                    returnType = ((BInvokableSymbol) symbolAtCursor).retType;
-                }
-                boolean hasError = false;
-                if (returnType instanceof BErrorType) {
-                    hasError = true;
-                } else if (returnType instanceof BUnionType) {
-                    BUnionType unionType = (BUnionType) returnType;
-                    hasError = unionType.getMemberTypes().stream().anyMatch(s -> s instanceof BErrorType);
-                    if (!isRemoteInvocation) {
-                        // Add type guard code action
-                        commandTitle = String.format(CommandConstants.TYPE_GUARD_TITLE, symbolAtCursor.name);
-                        List<TextEdit> tEdits = getTypeGuardCodeActionEdits(context, uri, refAtCursor, unionType);
-                        action = new CodeAction(commandTitle);
-                        action.setKind(CodeActionKind.QuickFix);
-                        action.setEdit(new WorkspaceEdit(Collections.singletonList(Either.forLeft(
-                                new TextDocumentEdit(new VersionedTextDocumentIdentifier(uri, null), tEdits)))));
-                        action.setDiagnostics(diagnostics);
-                        actions.add(action);
-                    }
-                }
-                // Add ignore return value code action
-                if (!hasError) {
-                    List<TextEdit> iEdits = getIgnoreCodeActionEdits(refAtCursor);
-                    commandTitle = CommandConstants.IGNORE_RETURN_TITLE;
-                    action = new CodeAction(commandTitle);
-                    action.setKind(CodeActionKind.QuickFix);
-                    action.setEdit(new WorkspaceEdit(Collections.singletonList(Either.forLeft(
-                            new TextDocumentEdit(new VersionedTextDocumentIdentifier(uri, null), iEdits)))));
-                    action.setDiagnostics(diagnostics);
-                    actions.add(action);
-                }
-            }
-        } catch (CompilationFailedException | WorkspaceDocumentException | IOException e) {
-            // ignore
-        }
-        return actions;
-    }
-
-    private static String getDiagnosedContent(Diagnostic diagnostic, LSContext context, LSDocumentIdentifier document) {
-        WorkspaceDocumentManager docManager = context.get(CodeActionKeys.DOCUMENT_MANAGER_KEY);
-        StringBuilder content = new StringBuilder();
-        Position start = diagnostic.getRange().getStart();
-        Position end = diagnostic.getRange().getEnd();
-        try (BufferedReader reader = new BufferedReader(
-                new StringReader(docManager.getFileContent(document.getPath())))) {
-            String strLine;
-            int count = 0;
-            while ((strLine = reader.readLine()) != null) {
-                if (count >= start.getLine() && count <= end.getLine()) {
-                    if (count == start.getLine()) {
-                        content.append(strLine.substring(start.getCharacter()));
-                        if (start.getLine() != end.getLine()) {
-                            content.append(System.lineSeparator());
-                        }
-                    } else if (count == end.getLine()) {
-                        content.append(strLine.substring(0, end.getCharacter()));
-                    } else {
-                        content.append(strLine).append(System.lineSeparator());
-                    }
-                }
-                if (count == end.getLine()) {
-                    break;
-                }
-                count++;
-            }
-        } catch (WorkspaceDocumentException | IOException e) {
-            // ignore error
-        }
-        return content.toString();
-    }
-
-    private static Position offsetInvocation(String diagnosedContent, Position position) {
-        // Diagnosed message only contains the erroneous part of the line
-        // Thus we offset into last
-        int leftParenthesisIndex = diagnosedContent.indexOf("(");
-        diagnosedContent = (leftParenthesisIndex == -1) ? diagnosedContent
-                : diagnosedContent.substring(0, leftParenthesisIndex);
-        String quotesRemoved = diagnosedContent
-                .replaceAll(".*:", "") // package invocation
-                .replaceAll(".*->", "") // action invocation
-                .replaceAll(".*\\.", ""); // object access invocation
-        int bal = diagnosedContent.length() - quotesRemoved.length();
-        if (bal > 0) {
-            position.setCharacter(position.getCharacter() + bal + 1);
-        }
-        return position;
-    }
-
-    private static List<TextEdit> getCreateVariableCodeActionEdits(LSContext context, String uri,
-                                                                   SymbolReferencesModel.Reference referenceAtCursor,
-                                                                   boolean hasDefaultInitFunction,
-                                                                   boolean hasCustomInitFunction) {
-        BLangNode bLangNode = referenceAtCursor.getbLangNode();
-        List<TextEdit> edits = new ArrayList<>();
-        CompilerContext compilerContext = context.get(DocumentServiceKeys.COMPILER_CONTEXT_KEY);
-        Set<String> nameEntries = CommonUtil.getAllNameEntries(bLangNode, compilerContext);
-        String variableName = CommonUtil.generateVariableName(bLangNode, nameEntries);
-
-        PackageID currentPkgId = bLangNode.pos.src.pkgID;
-        BiConsumer<String, String> importsAcceptor = (orgName, alias) -> {
-            boolean notFound = CommonUtil.getCurrentModuleImports(context).stream().noneMatch(
-                    pkg -> (pkg.orgName.value.equals(orgName) && pkg.alias.value.equals(alias))
-            );
-            if (notFound) {
-                String pkgName = orgName + "/" + alias;
-                edits.add(addPackage(pkgName, context));
-            }
-        };
-        String variableType;
-        if (hasDefaultInitFunction) {
-            BType bType = referenceAtCursor.getSymbol().type;
-            variableType = FunctionGenerator.generateTypeDefinition(importsAcceptor, currentPkgId, bType);
-            variableName = CommonUtil.generateVariableName(bType, nameEntries);
-        } else if (hasCustomInitFunction) {
-            BType bType = referenceAtCursor.getSymbol().owner.type;
-            variableType = FunctionGenerator.generateTypeDefinition(importsAcceptor, currentPkgId, bType);
-            variableName = CommonUtil.generateVariableName(bType, nameEntries);
-        } else {
-            // Recursively find parent, when it is an indexBasedAccessNode
-            while (bLangNode.parent instanceof IndexBasedAccessNode) {
-                bLangNode = bLangNode.parent;
-            }
-            variableType = FunctionGenerator.generateTypeDefinition(importsAcceptor, currentPkgId, bLangNode.type);
-        }
-        // Remove brackets of the unions
-        variableType = variableType.replaceAll("^\\((.*)\\)$", "$1");
-        String editText = createVariableDeclaration(variableName, variableType);
-        Position position = new Position(bLangNode.pos.sLine - 1, bLangNode.pos.sCol - 1);
-        edits.add(new TextEdit(new Range(position, position), editText));
-        return edits;
-    }
-
-    private static TextEdit addPackage(String pkgName, LSContext context) {
-        DiagnosticPos pos = null;
-
-        // Filter the imports except the runtime import
-        List<BLangImportPackage> imports = CommonUtil.getCurrentModuleImports(context);
-
-        if (!imports.isEmpty()) {
-            BLangImportPackage lastImport = CommonUtil.getLastItem(imports);
-            pos = lastImport.getPosition();
-        }
-
-        int endCol = 0;
-        int endLine = pos == null ? 0 : pos.getEndLine();
-
-        String editText = "import " + pkgName + ";\n";
-        Range range = new Range(new Position(endLine, endCol), new Position(endLine, endCol));
-        return new TextEdit(range, editText);
-    }
-
-    private static List<TextEdit> getTypeGuardCodeActionEdits(LSContext context, String uri,
-                                                              SymbolReferencesModel.Reference referenceAtCursor,
-                                                              BUnionType unionType)
-            throws WorkspaceDocumentException, IOException {
-        WorkspaceDocumentManager docManager = context.get(CodeActionKeys.DOCUMENT_MANAGER_KEY);
-        BLangNode bLangNode = referenceAtCursor.getbLangNode();
-        Position startPos = new Position(bLangNode.pos.sLine - 1, bLangNode.pos.sCol - 1);
-        Position endPosWithSemiColon = new Position(bLangNode.pos.eLine - 1, bLangNode.pos.eCol);
-        Position endPos = new Position(bLangNode.pos.eLine - 1, bLangNode.pos.eCol - 1);
-        Range newTextRange = new Range(startPos, endPosWithSemiColon);
-
-        List<TextEdit> edits = new ArrayList<>();
-        String spaces = StringUtils.repeat(' ', bLangNode.pos.sCol - 1);
-        String padding = LINE_SEPARATOR + LINE_SEPARATOR + spaces;
-        String content = getContentOfRange(docManager, uri, new Range(startPos, endPos));
-        boolean hasError = unionType.getMemberTypes().stream().anyMatch(s -> s instanceof BErrorType);
-
-        List<BType> members = new ArrayList<>((unionType).getMemberTypes());
-        long errorTypesCount = unionType.getMemberTypes().stream().filter(t -> t instanceof BErrorType).count();
-        boolean transitiveBinaryUnion = unionType.getMemberTypes().size() - errorTypesCount == 1;
-        if (transitiveBinaryUnion) {
-            members.removeIf(s -> s instanceof BErrorType);
-        }
-        // Check is binary union type with error type
-        if ((unionType.getMemberTypes().size() == 2 || transitiveBinaryUnion) && hasError) {
-            members.forEach(bType -> {
-                if (bType instanceof BNilType) {
-                    // if (foo() is error) {...}
-                    String newText = String.format("if (%s is error) {%s}", content, padding);
-                    edits.add(new TextEdit(newTextRange, newText));
-                } else {
-                    // if (foo() is int) {...} else {...}
-                    String type = CommonUtil.getBTypeName(bType, context, true);
-                    String newText = String.format("if (%s is %s) {%s} else {%s}",
-                                                   content, type, padding, padding);
-                    edits.add(new TextEdit(newTextRange, newText));
-                }
-            });
-        } else {
-            CompilerContext compilerContext = context.get(DocumentServiceKeys.COMPILER_CONTEXT_KEY);
-            Set<String> nameEntries = CommonUtil.getAllNameEntries(bLangNode, compilerContext);
-            String varName = CommonUtil.generateVariableName(bLangNode, nameEntries);
-            String typeDef = CommonUtil.getBTypeName(unionType, context, true);
-            boolean addErrorTypeAtEnd;
-
-            List<BType> tMembers = new ArrayList<>((unionType).getMemberTypes());
-            if (errorTypesCount > 1) {
-                tMembers.removeIf(s -> s instanceof BErrorType);
-                addErrorTypeAtEnd = true;
-            } else {
-                addErrorTypeAtEnd = false;
-            }
-            List<String> memberTypes = new ArrayList<>();
-            IntStream.range(0, tMembers.size())
-                    .forEachOrdered(value -> {
-                        BType bType = tMembers.get(value);
-                        String bTypeName = CommonUtil.getBTypeName(bType, context, true);
-                        boolean isErrorType = bType instanceof BErrorType;
-                        if (isErrorType && !addErrorTypeAtEnd) {
-                            memberTypes.add(bTypeName);
-                        } else if (!isErrorType) {
-                            memberTypes.add(bTypeName);
-                        }
-                    });
-
-            if (addErrorTypeAtEnd) {
-                memberTypes.add("error");
-            }
-
-            String newText = String.format("%s %s = %s;%s", typeDef, varName, content, LINE_SEPARATOR);
-            newText += spaces + IntStream.range(0, memberTypes.size() - 1)
-                    .mapToObj(value -> {
-                        return String.format("if (%s is %s) {%s}", varName, memberTypes.get(value), padding);
-                    })
-                    .collect(Collectors.joining(" else "));
-            newText += String.format(" else {%s}", padding);
-            edits.add(new TextEdit(newTextRange, newText));
-        }
-        return edits;
-    }
-
-    private static List<TextEdit> getIgnoreCodeActionEdits(SymbolReferencesModel.Reference referenceAtCursor) {
-        String editText = "_ = ";
-        BLangNode bLangNode = referenceAtCursor.getbLangNode();
-        Position position = new Position(bLangNode.pos.sLine - 1, bLangNode.pos.sCol - 1);
-        List<TextEdit> edits = new ArrayList<>();
-        edits.add(new TextEdit(new Range(position, position), editText));
-        return edits;
-
-    }
-
-    /**
-     * Inner class for the command argument holding argument key and argument value.
-     */
-    public static class CommandArgument {
-        private String argumentK;
-
-        private String argumentV;
-
-        public CommandArgument(String argumentK, String argumentV) {
-            this.argumentK = argumentK;
-            this.argumentV = argumentV;
-        }
-
-        public String getArgumentK() {
-            return argumentK;
-        }
-
-        public String getArgumentV() {
-            return argumentV;
-        }
-
     }
 }
