+++ language-server/modules/langserver-core/src/main/java/org/ballerinalang/langserver/command/executors/MessageExecutor.java
@@ -49,10 +49,14 @@ public class MessageExecutor implements LSCommandExecutor {
         for (Object arg : context.get(ExecuteCommandKeys.COMMAND_ARGUMENTS_KEY)) {
             String argKey = ((JsonObject) arg).get(ARG_KEY).getAsString();
             String argVal = ((JsonObject) arg).get(ARG_VALUE).getAsString();
-            if (argKey.equals(CommandConstants.ARG_KEY_MESSAGE_TYPE)) {
-                messageType = MessageType.forValue(Integer.parseInt(argVal));
-            } else if (argKey.equals(CommandConstants.ARG_KEY_MESSAGE)) {
-                message = argVal;
+            switch (argKey) {
+                case CommandConstants.ARG_KEY_MESSAGE_TYPE:
+                    messageType = MessageType.forValue(Integer.parseInt(argVal));
+                    break;
+                case CommandConstants.ARG_KEY_MESSAGE:
+                    message = argVal;
+                    break;
+                default:
             }
         }
         // If no package, or no doc uri; then just skip
