+++ language-server/modules/langserver-core/src/main/java/org/ballerinalang/langserver/common/constants/CommandConstants.java
@@ -29,7 +29,8 @@ public class CommandConstants {
     public static final String TAINTED_PARAM_PASSED = "tainted value passed to untainted parameter";
     public static final String NO_IMPL_FOUND_FOR_FUNCTION = "no implementation found for the function";
     public static final String FUNC_IMPL_FOUND_IN_ABSTRACT_OBJ = "cannot have a body";
-    public static final Pattern UNRESOLVED_MODULE_PATTERN = Pattern.compile("cannot resolve module '(.*)'");
+    public static final Pattern UNRESOLVED_MODULE_PATTERN = Pattern.compile(
+            "cannot resolve module '(.*)(?:version (.*)')|'");
     public static final Pattern TAINTED_PARAM_PATTERN = Pattern.compile(
             "tainted value passed to untainted parameter '(.*)'");
     public static final Pattern UNDEFINED_FUNCTION_PATTERN = Pattern.compile("undefined function '(.*)'");
